feat(signalStore): extract signal store as module
